refactor: move db.js into its own folder
